#2861 - Towards multi-value features
- Better factoring out of the ReorderableTagAutoCompleteField
- Formatting
