+++ inception/inception-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/feature/editor/KendoAutoCompleteTextFeatureEditor.java
@@ -20,8 +20,6 @@ package de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.editor;
 import static de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaBehavior.visibleWhen;
 import static org.apache.wicket.markup.head.JavaScriptHeaderItem.forReference;
 
-import java.util.List;
-
 import org.apache.wicket.MarkupContainer;
 import org.apache.wicket.markup.head.IHeaderResponse;
 import org.apache.wicket.markup.html.form.AbstractTextComponent;
@@ -31,7 +29,6 @@ import de.tudarmstadt.ukp.clarin.webanno.api.annotation.action.AnnotationActionH
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.keybindings.KeyBindingsPanel;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.FeatureState;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
-import de.tudarmstadt.ukp.clarin.webanno.model.ReorderableTag;
 
 /**
  * String feature editor using a Kendo AutoComplete field.
@@ -91,21 +88,6 @@ public class KendoAutoCompleteTextFeatureEditor
     @Override
     protected AbstractTextComponent createInputField()
     {
-        return new ReorderableTagAutoCompleteField("value")
-        {
-            private static final long serialVersionUID = 311286735004237737L;
-
-            @Override
-            protected List<ReorderableTag> getChoices(String aTerm)
-            {
-                FeatureState state = KendoAutoCompleteTextFeatureEditor.this.getModelObject();
-
-                TagRanker ranker = new TagRanker();
-                ranker.setMaxResults(maxResults);
-                ranker.setTagCreationAllowed(state.getFeature().getTagset().isCreateTag());
-
-                return ranker.rank(aTerm, state.tagset);
-            }
-        };
+        return new ReorderableTagAutoCompleteField("value", getModel(), maxResults);
     }
 }
