+++ inception/inception-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/feature/editor/StringFeatureTraitsEditor.html
@@ -24,7 +24,7 @@
         <div class="form-check">
           <input wicket:id="multipleRows" class="form-check-input" type="checkbox">
           <label wicket:for="multipleRows" class="form-check-label">
-            Multiple Rows
+            Multiple rows
           </label>
         </div>
       </div>
@@ -34,14 +34,14 @@
         <div class="form-check">
           <input wicket:id="dynamicSize" class="form-check-input" type="checkbox">
           <label wicket:for="dynamicSize" class="form-check-label">
-            Dynamic Size
+            Dynamic size
           </label>
         </div>
       </div>
     </div>
     <div class="row form-row" wicket:enclosure="collapsedRows">
       <label wicket:for="collapsedRows" class="col-sm-4 col-form-label">
-        Collapsed Rows
+        Collapsed rows
       </label>
       <div class="col-sm-8">
         <input type="number" style="width:100%;" wicket:id="collapsedRows" data-container="body"></input>
@@ -49,7 +49,7 @@
     </div>
     <div class="row form-row" wicket:enclosure="expandedRows">
       <label wicket:for="expandedRows" class="col-sm-4 col-form-label">
-        Expanded Rows
+        Expanded rows
       </label>
       <div class="col-sm-8">
         <input type="number" style="width:100%;" wicket:id="expandedRows" data-container="body"></input>
